fix admin

+++ backend/controllers/post.js
@@ -15,11 +15,8 @@ const getPosts = (req,res,next) => {
 const getByCategoriesId = (req,res, next) => {
   
     Post.find({
-        // category_id: req.params.categoryId,
-        // owner: req.user._id
-        category:{
-        category_id:req.params.categoryId
-        }
+         categoryId: req.params.categoryId,
+        //  owner: req.user._id
     })
     .then((posts) => res.send(posts))
     .catch(next)
@@ -32,41 +29,57 @@ const getPost = (req, res, next) => {
 }
 
 const createPost = (req, res, next) => {
-    const {title, seo_title, excerpt, body, slug, meta_description, meta_keywords, status, featured, views} = req.body;
-    const date = moment().format('DDMMYYYY-HHmmss_SSS')
-     let compressorImage = path.join(__dirname, '../', 'storage', '/', 'posts', req.file.originalname.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.jpg')
-   
-    sharp(req.file.path).resize(640).jpeg({
-        quality:80,
-        chromaSubsampling:'4:4:4'
-    }).toFile(compressorImage)
+    const {title, seo_title, excerpt, body, slug, meta_description, meta_keywords, status, featured, views, categoryId} = req.body;
+     let compressorImage
+   if (path.extname(req.file.originalname).toLowerCase() === ".png") {
+    compressorImage = path?.join(__dirname, '../', 'storage', '/', 'posts', req.file?.originalname?.split('.').slice(0, -1).join('.') + "-" + 'cropped' + '.png')
+   } else {
+    compressorImage = path?.join(__dirname, '../', 'storage', '/', 'posts', req.file?.originalname?.split('.').slice(0, -1).join('.') + "-" + 'cropped' + '.jpg')
+   }
 
-    Post.create({owner: req.user._id, category:req.body.category, image: req.file ? req.file.path : '', title, seo_title, excerpt, body, slug, meta_description, meta_keywords, status, featured, views})
+     if(req.file) {
+        sharp(req.file?.path).resize(640).jpeg({
+            quality:80,
+            chromaSubsampling:'4:4:4'
+        }).toFile(compressorImage)
+     }
+
+     console.log(compressorImage)
+
+    Post.create({owner: req.user._id, category:req.body.category, categoryId, image: req.file ? req.file.path : '', title, seo_title, excerpt, body, slug, meta_description, meta_keywords, status, featured, views})
     .then((post) => res.status(201).send(post))
     .catch(next)
-
-    
 }
-
 
 const deletePost = (req, res, next) => {
     Post.findById(req.params.id)
     .orFail(new NotFoundError('Пост не найден'))
     .then((post) => {
         if(req.user._id.toString() === post.owner.toString()) {
+            
             fs.unlink(post.image, (err) => {
                 if (err) {
                   console.error(err);
                 }
                 console.log("File removed");
               });
-    
-            fs.unlink(post.image.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.jpg', (err) => {
-                if (err) {
-                  console.error(err);
-                }
-                console.log("File removed");
-              });
+              if(post.image.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.jpg') {
+                fs.unlink(post.image.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.jpg', (err) => {
+                    if (err) {
+                      console.error(err);
+                    }
+                    console.log("File removed");
+                  });
+              }
+
+              if(post.image.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.png') {
+                fs.unlink(post.image.split('.').slice(0, -1).join('.') + "-" + 'cropped'+'.png', (err) => {
+                    if (err) {
+                      console.error(err);
+                    }
+                    console.log("File removed");
+                  });
+              }
              return post.remove()
             .then(() => res.send({message: "Пост удален"}))
         }
@@ -75,11 +88,10 @@ const deletePost = (req, res, next) => {
     .catch(next)
 }
 
-
-
 const updatePost = (req, res, next) => {
     const updated = {
         category:req.body.category,
+        categoryId: req.body.categoryId,
         title: req.body.title, 
         seo_title:req.body.seo_title, 
         except:req.body.except, 
@@ -110,7 +122,6 @@ const updatePost = (req, res, next) => {
     })
     .catch(next)
 }
-
 
 module.exports ={
     getByCategoriesId,
